requirements.txt
streamlit>=1.35.0
openai>=1.55.0
pypdf>=4.2.0
pdfplumber>=0.11.0
python-dotenv>=1.0.1
pytest>=8.3.0
